export send button as an atom component
